把表格移到查表格。
